python/requirements.txt
numpy
pandas
zat
matplotlib
python-Levenshtein 
